miniz: make it work on 64-bit Linux: http://en.wikipedia.org/wiki/64-bit#64-bit_data_models

+++ pteid-mw-pt/_src/eidmw/applayer/miniz.c
@@ -178,7 +178,7 @@ extern "C" {
 // ------------------- zlib-style API Definitions.
 
 // For more compatibility with zlib, miniz.c uses unsigned long for some parameters/struct members.
-typedef unsigned long mz_ulong;
+typedef unsigned int mz_ulong;
 
 // Heap allocation callbacks.
 // Note that mz_alloc_func parameter types purpsosely differ from zlib's: items/size is size_t, not unsigned long.
